% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{autoplot.sprite_df}
\alias{autoplot.sprite_df}
\title{Create a generic plot for sprite results}
\usage{
\method{autoplot}{sprite_df}(sprite_df, ...)
}
\arguments{
\item{sprite_df}{a dataframe that contains samples and answers from sprite.}

\item{...}{futher arguments to pass on to ggplot2}
}
\description{
This is a convenience function to quickly plot
the result of sprite. You can create and customize one yourself by using
ggplot(result,aes(answer))+geom_bar()+facet_wrap(~sample_id)
}
