% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grim.R
\name{check_mean_multi_item}
\alias{check_mean_multi_item}
\alias{check_mean}
\title{Granularity-Related Inconsistency of Means (GRIM)}
\source{
\url{https://osf.io/t47wf/}
The default options provided work with most cases
One item, 2 digits precision and integer scale (granularity =1)
So a mean of 15.25 with sample size (n) 4 is plausible
}
\usage{
check_mean_multi_item(mean, n, items = 1, digits = 2, G = 1)

check_mean(mean, n)
}
\arguments{
\item{mean}{mean to check}

\item{n}{number of observations}

\item{items}{how many observations}

\item{digits}{digits precision}

\item{G}{granularity}
}
\description{
This technique uses a very simple test to see if means are mathematically
possible  with a given sample size.
}
\details{
Technique first described by Nicholas J. L. Brown & James A. J. Heathers.
}
\section{Functions}{
\itemize{
\item \code{check_mean}: check_mean simple version
}}

\examples{
check_mean(mean = 20.54, n = 24)
}
\seealso{
Other granularity: 
\code{\link{check_sd_binary}()},
\code{\link{granularity}}
}
\concept{granularity}
