% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensicdatatoolkit-package.R
\docType{package}
\name{forensicdatatoolkit-package}
\alias{forensicdatatoolkit}
\alias{forensicdatatoolkit-package}
\title{forensicdatatoolkit: Forensic Data Toolkit}
\description{
This is a combination of tools for investigating data from the published literature.
      These tools do not need the underlying data, but use the reported summary statistics
      to detect anomalies.
}
\author{
\strong{Maintainer}: Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (\href{https://orcid.org/0000-0001-7509-0328}{ORCID})

Authors:
\itemize{
  \item Nicholas J. L. Brown (\href{https://orcid.org/0000-0003-1579-0730}{ORCID})
  \item James A.J.  Heathers (\href{https://orcid.org/0000-0002-0936-026X}{ORCID})
  \item Tim van der Zee (\href{https://orcid.org/000-0002-8058-9163}{ORCID})
  \item Jordan Anaya (\href{https://orcid.org/0000-0002-6166-4113}{ORCID})
}

}
\keyword{internal}
