% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granularity_checks.R
\name{granularity}
\alias{granularity}
\title{Granularity checks}
\description{
Most of the functions in this package make use of granularity checks on
summary statistics. Given a certain limits of sample size, lower and upper
scale and the fact that you can only answer with integers makes it
relatively easy to check for inconsistencies.
}
\details{
\itemize{
\item Check mean given certain n: \link[forensicdatatoolkit:check_mean]{check_mean()}
\item Check standard deviation given n and mean for binary data \link[forensicdatatoolkit:check_sd_binary]{check_sd_binary()}
}
}
\seealso{
Other granularity: 
\code{\link{check_mean_multi_item}()},
\code{\link{check_sd_binary}()}
}
\concept{granularity}
