% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grim.R
\name{grim_limit_n}
\alias{grim_limit_n}
\title{Grim limit}
\usage{
grim_limit_n(decimals, G = 1)
}
\description{
If the mean is reported to D decimal places
then the test can detect inconsistent means.
granularity / N > 10e-D
In other words, using only integers (G=1)
and 2 decimals, what is the limit where
grim is no longer able to find issues?
100, so up to and including 99 it works fine.
}
