% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grim.R
\name{grim_multi_item}
\alias{grim_multi_item}
\alias{grim}
\title{Granularity-Related Inconsistency of Means (GRIM)}
\source{
\url{https://osf.io/t47wf/}
The default options provided work with most cases
One item, 2 digits precision and integer scale (granularity =1)
So a mean of 15.25 with sample size (n) 4 is plausible
}
\usage{
grim_multi_item(mean, n, items = 1, digits = 2, G = 1)

grim(mean, n)
}
\arguments{
\item{mean}{mean to check}

\item{n}{number of observations}

\item{items}{how many observations}

\item{digits}{digits precision}

\item{G}{granularity}
}
\description{
This technique uses a very simple test to see if means are mathematically
possible  with a given sample size.
}
\details{
Technique first described by Nicholas J. L. Brown & James A. J. Heathers.
}
\section{Functions}{
\itemize{
\item \code{grim}: Grim simple version
}}

\examples{
grim(mean = 20.54, n = 24)
}
