% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{make_mean_grim}
\alias{make_mean_grim}
\title{Copy from script, slight modifications to keep argument names consistent.}
\usage{
make_mean_grim(mean, n, dp = 2)
}
\description{
Copy from script, slight modifications to keep argument names consistent.
}
