% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{n_unique_rows}
\alias{n_unique_rows}
\title{Count unique results}
\usage{
n_unique_rows(matrix)
}
\arguments{
\item{matrix}{to count the rows of}
}
\value{
number.
}
\description{
Count unique results
}
