% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{non_zero_rows}
\alias{non_zero_rows}
\title{Find rows that are not zero
Which we began with.}
\usage{
non_zero_rows(matrix)
}
\arguments{
\item{matrix}{to count the rows of}
}
\value{
logical vector of rows that are not zero
}
\description{
Find rows that are not zero
Which we began with.
}
