% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{not_duplicate_rows}
\alias{not_duplicate_rows}
\title{Find non duplicate rows in a matrix}
\usage{
not_duplicate_rows(matrix)
}
\arguments{
\item{matrix}{to count the rows of}
}
\value{
logical vector of rows that are not duplicates
}
\description{
Find non duplicate rows in a matrix
}
