% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{pick_one_position}
\alias{pick_one_position}
\title{select a random TRUE position out of logical vector}
\usage{
pick_one_position(logical_vector)
}
\description{
select a random TRUE position out of logical vector
}
