% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite.R
\name{simulate_samples}
\alias{simulate_samples}
\title{Get sprite samples data.frame}
\usage{
simulate_samples(
  max_cases,
  n,
  mean,
  sd,
  scale_min,
  scale_max,
  dp = 2,
  fixed = c()
)
}
\arguments{
\item{max_cases}{maximum cases}

\item{n}{total cases}

\item{mean}{mean value}

\item{sd}{standard deviation}

\item{scale_min}{minimum of scale}

\item{scale_max}{maximum of scale}

\item{dp}{degrees of precision, places behind the comma}

\item{fixed}{a vector of values that you know are in the sprite sample}
}
\description{
Create samples that match the mean and sd.
}
\details{
\subsection{Advanced use}{

It is also possible to use the underlying functions directly.
\itemize{
\item \link[forensicdatatoolkit:simulate_vector]{simulate_vector()} if you need a single vector only
\item \link[forensicdatatoolkit:simulate_samples_matrix]{simulate_samples_matrix()} if you want to work with a matrix of vectors
\item if you would like to turn the matrix into a dataframe use the \link[forensicdatatoolkit:sprite_into_df]{sprite_into_df()} function
}
}
}
\examples{
simulate_samples(10, 10, 2.5, 2, 1, 5)
simulate_samples(2, 20, 2.5, 2, 1, 5, fixed = c(1, 1))
}
\seealso{
Other sprite: 
\code{\link{simulate_samples_matrix}()},
\code{\link{simulate_vector}()},
\code{\link{sprite_into_df}()}
}
\concept{sprite}
