% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite.R
\name{simulate_vector}
\alias{simulate_vector}
\title{Find single vector of mean and sd.}
\usage{
simulate_vector(mean, sd, n, scale_min, scale_max, dp = 2, fixed = c())
}
\arguments{
\item{mean}{mean value}

\item{sd}{standard deviation}

\item{n}{total cases}

\item{scale_min}{minimum of scale}

\item{scale_max}{maximum of scale}

\item{dp}{degrees of precision, places behind the comma}

\item{fixed}{a vector of values that you know are in the sprite sample}
}
\description{
Return a vector of length n with an average close to mean and
standard deviation close to sd. In most cases you want more than one
sample, to see what are feasable solutions see \link[forensicdatatoolkit:sprite_samples]{sprite_samples()}
}
\seealso{
Other sprite: 
\code{\link{simulate_samples_matrix}()},
\code{\link{simulate_samples}()},
\code{\link{sprite_into_df}()}
}
\concept{sprite}
