% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{sprite_into_df}
\alias{sprite_into_df}
\title{Aggregate sprite results into data.frame}
\usage{
sprite_into_df(sprite_matrix)
}
\description{
Convenience function for sprite so you can easily work
with the results in further work.
Sprite returns a matrix with a result per row.
This is computational efficient but not easy to work with.
}
\seealso{
Other sprite: 
\code{\link{simulate_samples_matrix}()},
\code{\link{simulate_samples}()},
\code{\link{simulate_vector}()}
}
\concept{sprite}
