% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{step_size}
\alias{step_size}
\title{Function that outputs 1 most of the times but <20\% of the time 2}
\usage{
step_size()
}
\description{
Function that outputs 1 most of the times but <20\% of the time 2
}
