% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprite_helpers.R
\name{unique_rows}
\alias{unique_rows}
\title{All unique columns
combine non zero and not duplicate cols}
\usage{
unique_rows(matrix)
}
\arguments{
\item{matrix}{to count the rows of}
}
\value{
a logical vector indicating unique columns in a matrix
}
\description{
All unique columns
combine non zero and not duplicate cols
}
